requirements.txt
firebase-admin=6.6.0
python-dotenv=1.0.1